属性変更機能の廃止 クリスタルが淡く光るように変更 それにともないほぼすべてのオブジェクトにシェーダーを適応

+++ Scenes/PlayScene/Field/FieldObjectManager.cpp
@@ -72,7 +72,7 @@ void FieldObjectManager::Initialize()
 
 		SimpleMath::Vector3 crystalPos = SimpleMath::Vector3(loadCrysralPos.x,-1.0f, loadCrysralPos.y);
 
-		m_crystals->push_back(*std::make_unique<Crystal>(crystalPos, cos(5)));
+		m_crystals->push_back(*std::make_unique<Crystal>(crystalPos, (float)cos(5)));
 	}
 
 	// ƒNƒŠƒXƒ^ƒ‹‚Ì‰Šú‰»ˆ—
